added deploy-command
