+++ deploy-commands.js
@@ -1,0 +1,49 @@
+import { REST,Routes } from "discord.js";
+import { config } from 'dotenv';
+import * as fs from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'url';
+
+config();
+const commands = [];
+const TOKEN = process.env.DISCORD_TOKEN;
+const CLIENT_ID = process.env.NODE_DISCORD_APP_ID;
+const GUILD_ID = process.env.SERVER_ID
+
+// Conseguir archivos de comandos y cargarlos
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const commandsPath = path.join(__dirname,'commands')
+const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('.js'));
+
+for (const file of commandFiles) {
+    const filePath = path.join(commandsPath,file);
+    const { default: command } = await import(filePath);
+    // Set a new item in the Collection with the key as the command name and the value as the exported module
+    if ('data' in command && 'execute' in command) {
+        commands.push(command.data.toJSON());
+    } else {
+        console.log(`[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`);
+    }
+}
+
+// Construct and prepare an instance of the REST module
+const rest = new REST().setToken(TOKEN);
+
+(async () => {
+    try {
+        console.log(`Started refreshing ${commands.length} application (/) commands.`);
+
+        // The put method is used to fully refresh all commands in the guild with the current set
+        const data = await rest.put(
+            Routes.applicationGuildCommands(CLIENT_ID,GUILD_ID),
+            { body: commands },
+        );
+
+        console.log(`Successfully reloaded ${data.length} application (/) commands.`);
+    } catch (error) {
+        // And of course, make sure you catch and log any errors!
+        console.error(error);
+    }
+})();
+
