+++ index.js
@@ -1,6 +1,5 @@
 import { Client,Collection,Events,GatewayIntentBits,REST } from 'discord.js';
 import { config } from 'dotenv';
-import * as fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'url';
 
@@ -24,44 +23,6 @@ const client = new Client({
 
 client.commands = new Collection();
 
-const commandsPath = path.join(__dirname,'commands')
-const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('.js'));
-
-for (const file of commandFiles) {
-	const filePath = path.join(commandsPath,file);
-	const { default: command } = await import(filePath);
-	console.log(filePath);
-	// Set a new item in the Collection with the key as the command name and the value as the exported module
-	if ('data' in command && 'execute' in command) {
-		client.commands.set(command.data.name,command);
-	} else {
-		console.log(`[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`);
-	}
-}
-
-
-
-// try {
-// 	console.log('Started refreshing application (/) commands.');
-
-// 	await rest.put(Routes.applicationCommands(CLIENT_ID),{ body: commands });
-
-// 	console.log('Successfully reloaded application (/) commands.');
-// } catch (error) {
-// 	console.error(error);
-// }
-
-// handle slash commands
-// client.on('interactionCreate',async interaction => {
-// 	if (!interaction.isChatInputCommand()) return;
-
-// 	if (interaction.commandName === 'ping') {
-// 		await interaction.reply('Pong!');
-// 	} else if (interaction.commandName === 'warning') {
-// 		console.log(interaction.options.get('user'))
-// 		await interaction.reply(`ahora ${interaction.options.get('user').user} se gano un warning por: ${interaction.options.get('motivo').value}`)
-// 	}
-// });
 
 client.on(Events.InteractionCreate,async interaction => {
 	if (!interaction.isChatInputCommand()) return;
