+++ package.json
@@ -27,6 +27,7 @@
 		"@discordjs/core": "^1.0.1",
 		"@discordjs/rest": "^2.0.1",
 		"@discordjs/ws": "^1.0.1",
+		"discord.js": "^14.13.0",
 		"dotenv": "^16.3.1",
 		"node": "^20.6.0",
 		"nodemon": "^3.0.1"
